requirements.txt
wait-for-it==2.2.0
pytest==6.2.5
pre-commit==2.15.0
python-dotenv==1.0.0
pydantic==1.10.7
redis==4.5.5
orjson==3.8.10
httptools==0.5.0
gunicorn==20.1.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
backoff==2.2.1
flask==2.3.2
flask-marshmallow==0.15.0
flask_sqlalchemy==3.0.3
flask-jwt-extended==4.4.4
requests==2.31.0
gevent==22.10.2
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
passlib==1.7.4
flask_jwt_extended==4.4.4
marshmallow-sqlalchemy==0.29.0
flask-migrate==4.0.4
opentelemetry-api==1.10.0
opentelemetry-sdk==1.10.0
opentelemetry-instrumentation-flask==0.29b1
opentelemetry-exporter-jaeger==1.10.0
Flask-Limiter==3.3.1
Authlib==1.2.0
python-logstash==0.4.8
flask_cors==4.0.0
yfinance==0.2.37
Flask-APScheduler==1.13.1
matplotlib==3.8.3
numpy==1.26.4
plotly==5.19.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash==2.16.1
